Fix parameter type ordering
Until now, the parameter types were obtained by calling `.values()` on
the HashSet of parameter types, but this leads to unpredicable ordering,
causing queries to fail because their parameters were often transposed.

Instead, sort the parameter types by their names.  This should always
work because in the postgresql case, the keys are '$1', '$2', etc.